pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "notion-query"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.5"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chonkie>=1.5.6",
    "fastembed>=0.7.4",
    "json-repair>=0.58.6",
    "langchain-core>=1.2.17",
    "langgraph>=1.0.10",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "langgraph-cli[inmem]>=0.4.14",
    "langsmith>=0.7.14",
    "matplotlib>=3.10.8",
    "nest-asyncio>=1.6.0",
    "notion-client>=3.0.0",
    "numpy>=2.4.3",
    "openai>=2.26.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "pyperclip>=1.9.0",
    "pyyaml>=6.0.3",
    "qdrant-client>=1.17.0",
    "requests>=2.32.5",
    "typer>=0.24.1",
    "pytest-recording>=0.13.4",
    "pytest-cov>=7.1.0",
    "pytest-asyncio>=1.3.0",
    "rich>=14.3.3",
]

[project.scripts]
notion-agent = "notion_query.cli:app"

[tool.uv]
package = true

[tool.setuptools.packages.find]
where = ["."]
include = ["notion_query*", "src*"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
